requirements.txt
asgiref==3.3.4
astroid==2.5.6
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.2.2
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.1.2
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pylint==2.8.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
